Add more warnings to MSVC and fix them

+++ src/figure/formation_herd.c
@@ -29,7 +29,7 @@
 // Look for a free tile, in the neighborhood of (x,y)
 static int get_free_tile(int x, int y, int allow_negative_desirability, int *x_tile, int *y_tile)
 {
-    unsigned short disallowed_terrain = ~(TERRAIN_ACCESS_RAMP | TERRAIN_MEADOW);
+    int disallowed_terrain = ~(TERRAIN_ACCESS_RAMP | TERRAIN_MEADOW);
     int tile_found = 0;
     int x_found = 0, y_found = 0;
